% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_field.R
\name{clean_field}
\alias{clean_field}
\title{Clean results by fields.}
\usage{
clean_field(data, ...)
}
\arguments{
\item{data}{a data frame object.}

\item{...}{arguments specifying fields with or without values, in the format of \code{'field'} or \code{'field'='value'}. Separate multiple arguments with comma.}
}
\value{
cleaned data frame or \code{FALSE} if error occurs or cleaned result has 0 row.
}
\description{
Removes all records in a libgenr object that don't have specified field(s) or field(s) values.
}
\examples{
clean_field(data=libgen_search('statistics'),'year'='2006','author')
clean_field(data=libgen_new_uploads(timefirst=Sys.Date()),'language'='English')

}
