% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_field.R
\name{has_field}
\alias{has_field}
\title{Tell if a certain field is contained.}
\arguments{
\item{field}{the field to look for.}

\item{data}{a data frame object.}
}
\value{
a logical value.
}
\description{
This function tells if an object returned from \code{libgen_retrieve_ID()}, \code{libgen_search()} or \code{libgen_new_uploads()} contains a certain field.
}
\examples{
has_field('Title',libgen_ID_retrieve(c(1,2,3)))

}
