% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libgen_download.R
\name{libgen_download}
\alias{libgen_download}
\title{Download books.}
\usage{
libgen_download(md5)
}
\arguments{
\item{md5}{The unique md5 hash code of a book from Library Genesis API.}
}
\description{
This function downloads books using md5 hash code. The download path is current working directory.
}
\examples{
libgen_download('30392A7F10E7B1864670820F2390713A')
}
