% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libgen_ID_retrieve.R
\name{libgen_ID_retrieve}
\alias{libgen_ID_retrieve}
\title{Get books information by IDs.}
\usage{
libgen_ID_retrieve(IDs, fields = c(), full_fields = FALSE)
}
\arguments{
\item{IDs}{One book ID or a vector of book IDs.}

\item{fields}{Optional, user specified vector of fields to return. Below is available fields (case sensitive):
ID, Title, VolumeInfo, Series, Periodical, Author,Year, Edition, Publisher,
City, Pages, Language, Topic,Library, Issue, Identifier, ISSN, ASIN, UDC, LBC, DDC,
LCC, Doi, Googlebookid, OpenLibraryID, Commentary, DPI,
Color, Cleaned, Orientation, Paginated, Scanned,
Bookmarked, Searchable,Filesize,  Extension, MD5, CRC32,
eDonkey, AICH, SHA1, TTH, Generic,  Visible, Locator,
Local, TimeAdded, TimeLastModified, Coverurl, identifierwodash,
tags, pagesinfile, descr, toc, tth, sha1, aich, btih, torrent, crc32}

\item{full_fields}{Whether returns with full fields (\code{TRUE}) or selected fields (\code{FALSE}).}
}
\value{
A data frame with results or FALSE if error occurs.
}
\description{
This function gets books information by IDs with user specified or default fields.
}
\examples{
libgen_ID_retrieve(c(1,2,3))
libgen_ID_retrieve(c(1,2,3),full_fields=FALSE)
libgen_ID_retrieve(c(1,2,3),fields=c('ID','Title','Author'))
}
