% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libgen_search.R
\name{libgen_search}
\alias{libgen_search}
\title{Search the Library Genesis API by text}
\usage{
libgen_search(query, count = 25, search_in = "def", sort_by = "def",
  reverse = FALSE, full_fields = FALSE, fields = c())
}
\arguments{
\item{query}{the string to search for.}

\item{count}{the number of results to return, defaults to 25. If the results length is less than \code{count}, will return the whole results.}

\item{search_in}{restrict your search to one of the following fields:
  \itemize{
 \item title
 \item author
 \item series
 \item periodical
 \item publisher
 \item year
 \item identifier
 \item md5
 \item extension
 \item def (default:Title, Author, Series, Periodical, Publisher, Year, VolumeInfo)
}}

\item{sort_by}{The field by which results are sorted:
\itemize{
 \item title
 \item author
 \item publisher
 \item year
 \item pages
 \item language
 \item extension
 \item filesize
 \item extension
 \item def (default: Relevance)
}}

\item{reverse}{Only if sort_by field is specified. Default to \code{FALSE}, if \code{TRUE}, sorts Z-A or 9-0.
Notice: if results are sorted by a field and \code{reverse} set to \code{FALSE}, items with no record on that field will come first.}

\item{full_fields}{whether returns with full fields (\code{TRUE}) or selected fields (\code{FALSE}).}

\item{fields}{Optional, user specified vector of fields to return.}
}
\value{
A data frame.
}
\description{
Search the Library Genesis API by text and return results of specified counts and sorted on specified fields.
}
\examples{
libgen_search('statistical learning')
libgen_search('statistical learning',count=150,search_in='title')
libgen_search('statistical learning',sort_by='year',reverse=TRUE)
}
